REFACTOR: Board and enum

+++ server/src/main/java/ch/heigvd/DIRECTION.java
@@ -1,11 +1,17 @@
 package ch.heigvd;
 
+/**
+ * The enum that represent the position on the map
+ */
 public enum DIRECTION {
     UP,
     DOWN,
     LEFT,
     RIGHT;
 
+    /**
+     * The parser that convert a key to a direction
+     */
     public static DIRECTION parseKey(KEY key){
 
         switch (key) {
@@ -27,13 +33,19 @@ public enum DIRECTION {
 
         }
     }
-    public static short[] getCoef(DIRECTION direction){
+
+    /**
+     * Get the coef of the direction
+     * @param direction The direction to get the coef
+     * @return The coef of the direction
+     */
+    public static int[] getCoef(DIRECTION direction){
         return switch (direction) {
-            case UP -> new short[]{0, -1};
-            case DOWN -> new short[]{0, 1};
-            case LEFT -> new short[]{-1, 0};
-            case RIGHT -> new short[]{1, 0};
-            default -> new short[]{0, 0};
+            case UP -> new int[]{0, -1};
+            case DOWN -> new int[]{0, 1};
+            case LEFT -> new int[]{-1, 0};
+            case RIGHT -> new int[]{1, 0};
+            default -> new int[]{0, 0};
         };
     }
 }
